refactor(webapp_v2): drop pages/Features wrapper and /features route prefix
The Features/ subfolder was a holdover from the CLJS organization where
"features" meant a specific cluster of feature-flagged pages. In the
React app every page is conceptually a feature, so the extra level only
adds noise.

Moved pages/Features/AiAgentsIdentities/ up one level to
pages/AiAgentsIdentities/ and removed the /features prefix from its
route — now /ai-agents-identities. Sidebar entry and the page's own
LIST_PATH / NEW_PATH / CREATED_PATH constants updated accordingly.

The other Features/ entries (AccessControl, DataMasking, Runbooks)
were unwired one-line stubs that nothing imported — deleted along with
the now-empty Features/ directory. CLJS routes under /features/* are
untouched: this change is React-only, and CLJS pages keep their paths
until they are individually migrated.
