--- webapp_v2/src/pages/Features/AccessControl/Create/index.jsx
@@ -1,7 +1,0 @@
-import { Title } from '@mantine/core'
-
-function AccessControlCreate() {
-  return <Title order={2}>Create Access Control Rule</Title>
-}
-
-export default AccessControlCreate
