--- webapp_v2/src/pages/Features/AccessControl/index.jsx
@@ -1,7 +1,0 @@
-import { Title } from '@mantine/core'
-
-function AccessControl() {
-  return <Title order={2}>Access Control</Title>
-}
-
-export default AccessControl
