--- webapp_v2/src/pages/Features/DataMasking/Create/index.jsx
@@ -1,7 +1,0 @@
-import { Title } from '@mantine/core'
-
-function DataMaskingCreate() {
-  return <Title order={2}>Create Data Masking Rule</Title>
-}
-
-export default DataMaskingCreate
