--- webapp_v2/src/pages/Features/DataMasking/index.jsx
@@ -1,7 +1,0 @@
-import { Title } from '@mantine/core'
-
-function DataMasking() {
-  return <Title order={2}>Data Masking</Title>
-}
-
-export default DataMasking
