--- webapp_v2/src/pages/Features/Runbooks/Setup/index.jsx
@@ -1,7 +1,0 @@
-import { Title } from '@mantine/core'
-
-function RunbooksSetup() {
-  return <Title order={2}>Runbooks Setup</Title>
-}
-
-export default RunbooksSetup
