--- webapp_v2/src/pages/Features/Runbooks/index.jsx
@@ -1,7 +1,0 @@
-import { Title } from '@mantine/core'
-
-function Runbooks() {
-  return <Title order={2}>Runbooks</Title>
-}
-
-export default Runbooks
